+++ dash_labs/plugins/__init__.py
@@ -1,6 +1,5 @@
 from .flexible_callbacks import FlexibleCallbacks
 from .hidden_components import HiddenComponents
 from .long_callback import LongCallback
-from .long_callback.long_callback_celery import LongCallback2
-from .long_callback.backends.celery_backend import CeleryCallbackManager
-from .long_callback.backends.flask_caching_backend import FlaskCachingCallbackManager
+from .long_callback.managers.celery_manager import CeleryCallbackManager
+from .long_callback.managers.flask_caching_manager import FlaskCachingCallbackManager
