+++ dash_labs/plugins/long_callback/__init__.py
@@ -2,18 +2,16 @@ from functools import partial
 from types import MethodType
 import uuid
 import json
-from multiprocessing import Process
 
 
 class LongCallback:
-    def __init__(self, cache):
-        self.cache = cache
+    def __init__(self, long_callback_manager):
+        self.long_callback_manager = long_callback_manager
 
     def plug(self, app):
-        self.cache.init_app(app.server)
+        self.long_callback_manager.init(app)
         app._background_procs = {}
-        app._flask_caching_cache = self.cache
-
+        app._long_callback_manager = self.long_callback_manager
         app.long_callback = MethodType(partial(long_callback), app)
 
 
@@ -36,14 +34,16 @@ def long_callback(self, args, output, running=(), cancel=(), progress=(), templa
 
     cancel_prop_ids = tuple(".".join([dep.component_id, dep.component_property]) for dep in cancel)
 
-    # Initialize procs list on app
-    procs = self._background_procs
+    callback_manager = self._long_callback_manager
 
     # Add hidden components to app so we don't need to put them in the layout
     hidden_components = self._extra_hidden_components
     hidden_components.extend([interval_component, store_component])
 
     def wrapper(fn):
+
+        background_fn = callback_manager.make_background_fn(fn, progress=bool(progress))
+
         def callback(n_intervals, cancel, user_store_data, user_callback_args):
             if isinstance(user_callback_args, tuple):
                 key = [list(user_callback_args), user_id]
@@ -51,18 +51,17 @@ def long_callback(self, args, output, running=(), cancel=(), progress=(), templa
                 key = [user_callback_args, user_id]
 
             result_key = json.dumps({"args": key})
-            progress_key = result_key + "-progress"
 
             should_cancel = any(
                 [trigger["prop_id"] in cancel_prop_ids for trigger in dash.callback_context.triggered]
             )
-            no_update_progress = map_grouping(
-                        lambda x: dash.no_update, progress.dependencies()
+            clear_progress = map_grouping(
+                        lambda x: None, progress.dependencies()
             ) if progress else ()
+
             if should_cancel:
-                if result_key in procs:
-                    for proc in procs.pop(result_key):
-                        proc.kill()
+                if callback_manager.has_future(result_key):
+                    callback_manager.cancel_future(result_key)
                 return dict(
                     user_callback_output=map_grouping(
                         lambda x: dash.no_update, output.dependencies()
@@ -70,28 +69,23 @@ def long_callback(self, args, output, running=(), cancel=(), progress=(), templa
                     in_progress=tuple([
                         val for (_, _, val) in running
                     ]),
-                    progress=no_update_progress,
+                    progress=clear_progress,
                     interval_disabled=True
                 )
 
-            result = self._flask_caching_cache.get(result_key)
-            progress_result = self._flask_caching_cache.get(progress_key)
+            progress_tuple = callback_manager.get_progress(result_key)
 
-            if result is not None:
-                if result_key in procs:
-                    for proc in procs.pop(result_key):
-                        proc.join()
-                        proc.close()
-
+            if callback_manager.result_ready(result_key):
+                result = callback_manager.get_result(result_key)
                 return dict(
                     user_callback_output=result,
                     in_progress=tuple([
                         val for (_, _, val) in running
                     ]),
-                    progress=progress_result if progress_result else no_update_progress,
+                    progress=clear_progress,
                     interval_disabled=True,
                 )
-            elif progress_result:
+            elif progress_tuple:
                 return dict(
                     user_callback_output=map_grouping(
                         lambda x: dash.no_update, output.dependencies()
@@ -99,16 +93,13 @@ def long_callback(self, args, output, running=(), cancel=(), progress=(), templa
                     in_progress=tuple([
                         val for (_, val, _) in running
                     ]),
-                    progress=progress_result,
+                    progress=progress_tuple,
                     interval_disabled=False,
                 )
             else:
-                target = make_update_cache(
-                    fn, self._flask_caching_cache, result_key, progress_key if progress else None
-                )
-                p = Process(target=target, args=(user_callback_args,))
-                p.start()
-                procs.setdefault(result_key, []).append(p)
+                callback_manager.terminate_unhealthy_future(result_key)
+                if not callback_manager.has_future(result_key):
+                    callback_manager.call_and_register_background_fn(result_key, background_fn, user_callback_args)
 
                 return dict(
                     user_callback_output=map_grouping(
@@ -117,7 +108,7 @@ def long_callback(self, args, output, running=(), cancel=(), progress=(), templa
                     in_progress=tuple([
                         val for (_, val, _) in running
                     ]),
-                    progress=no_update_progress,
+                    progress=clear_progress,
                     interval_disabled=False,
                 )
 
@@ -142,20 +133,3 @@ def long_callback(self, args, output, running=(), cancel=(), progress=(), templa
         )(callback)
 
     return wrapper
-
-
-def make_update_cache(fn, cache, result_key, progress_key=None):
-    def _set_progress(i, total):
-        cache.set(progress_key, (i, total))
-
-    def _callback(user_callback_args):
-        maybe_progress = [_set_progress] if progress_key is not None else []
-        if isinstance(user_callback_args, dict):
-            user_callback_output = fn(*maybe_progress, **user_callback_args)
-        elif isinstance(user_callback_args, list):
-            user_callback_output = fn(*maybe_progress, *user_callback_args)
-        else:
-            user_callback_output = fn(*maybe_progress, user_callback_args)
-        cache.set(result_key, user_callback_output)
-
-    return _callback
