--- dash_labs/plugins/long_callback/backends/__init__.py
@@ -1,11 +1,0 @@
-from abc import ABC
-
-class LongCallbackBackend(ABC):
-
-    def send_result(self, value):
-        pass
-
-    def send_progress(self, progress, total):
-        pass
-
-
