+++ docs/demos/long_callback_examples/long_callback_celery_loading_state.py
@@ -4,12 +4,24 @@ import dash
 import dash_core_components as dcc
 import dash_html_components as html
 import dash_labs as dl
+from celery import Celery
+
 from flask_caching import Cache
 
-cache = Cache(config={"CACHE_TYPE": "filesystem", "CACHE_DIR": "./cache"})
+from dash_labs.plugins import FlaskCachingCallbackManager, CeleryCallbackManager
+
+# celery_app = Celery(__name__, backend='rpc://', broker='pyamqp://')
+# long_callback_manager = CeleryCallbackManager(celery_app)
+
+flask_cache = Cache(config={"CACHE_TYPE": "filesystem", "CACHE_DIR": "./cache"})
+long_callback_manager = FlaskCachingCallbackManager(flask_cache)
+
 app = dash.Dash(__name__, plugins=[
-    dl.plugins.FlexibleCallbacks(), dl.plugins.HiddenComponents(), dl.plugins.LongCallback(cache)
+    dl.plugins.FlexibleCallbacks(),
+    dl.plugins.HiddenComponents(),
+    dl.plugins.LongCallback(long_callback_manager)
 ])
+
 
 app.layout = html.Div([
     html.Div([
@@ -30,10 +42,9 @@ app.layout = html.Div([
         (dl.Output("cancel_button_id", "disabled"), False, True),
     ],
     cancel=[dl.Input("cancel_button_id", "n_clicks")],
-    interval=300,
 )
 def callback(n_clicks):
-    time.sleep(2.0)
+    time.sleep(5.0)
     return [f"Clicked {n_clicks} times"]
